+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add more local data source attributes
 - Implement Campusonline Public REST API, caching the results in a local database table and getting organizations form there
 
 ## v0.1.23